refactor: fix yaml_utils naming typo and implement ThemeManager
