+++ src/pylunix/utils/__init__.py
@@ -1,5 +1,5 @@
 from .qss_utils import QSSProcessor
-from .yaml_util import YAMLProcessor
+from .yaml_utils import YAMLProcessor
 from .math_utils import safe_eval_math, IncrementNumberRounder, DecimalFormatter
 from .string_utils import extract_numbers
